changes to get a working blast submission w/ successful retrieval
how the blast submission works is that a job is started for all the query sequences.
and then, the table is split into separate dataframes by the queries (often there is one query so the split does not matter).
the second job is batched into an array with a length equal to the number of splits (1 in the case of a single query).

each split gets its own output sub-directory and there is an output sub-directory for the query data only as well.

a lot of the code previous code had comments that conflicted the true state, making things extremely difficult to understand and debug.
now, all the comments should be updated, but please leave a github comment if something is not clear.

there's still a few outdated comments in the other wrapper scripts that would be worth changing.

i initially thought the sequences were being split by homolog, not by query. so the initial changes i made on this branch would spawn 1
interproscan process per homolog which would blow up the memory because of how the iprscan command line tool allocates cpu and memory per process.

by utilizing multifastas instead as the input for the iprscan process n_query_proteins + 1 processes are created now (which is typically only 2 total);
instead of the initial commit where the number of processes spawned was equal to the number of homologs