fix and verify retrieval works for all examples, cleanup & more
- after spending a lot more time testing on the dev env,
i've found a few small fixes to have job retrieval work for a few types that weren't
before: fasta/accnum/msa phylogenetic analysis, ipr without homology search, & more.

most of these fixes involved chceking if a file is created an using that to determine
the number of jobs. phylogenetic and non homology search jobs only run analysis on the
queries (1 job). while for homology search jobs, we batch the jobs and use the number
of records in the logfile to count.

now, this is still prone to the issue of a job failing, but at least the retrieval system
shouldn't be at fault now; although, there's a chance i missed a few edge cases since i mostly
used the built-in examples plus a few i had ready locally

- logfiles now write NA for missing values and are set using += string operator to allow line wrapping

a missing value occurs very often for the query_data job row where some analysis setps are skipped.

- add a comment explaining why we overwrite the AccNum column for blast submissions

- add parameter validation for wrapper scripts (only for parameters that are required)

- update Lauren -> Lo

- rm the helper functions for now. they are setup to be sourced in the dev env (i.e., mounted in container),
but i don't think they are necessary to use for now. at least the idea and framework is there if want to
include bash functions

- other misc minor comment updates such as updating outdated usage statements and giving
an explanation on how some of the job retrieval checking works