document and prevent warnings from summarize functions
I was documenting the code to understand it. Now, I'm pushing the documentation

- One minor change is that `select` was being used before in
`elements2words` which lead to strange resulting string of
'c( w1 w2 w-3)'. Now with pull, the output is much better: 'w1 w2 w-3'.

Because the return value is always fed into words2wc, this fix is not
essential but is better, imo.

- Add Roxygen2 documentation to a few functions